This test normally takes about **500s** in Travis and sometimes times out. Locally, this PR speeds up the test for me from **140s** to **30s**. On Travis it speeds it up from **500s** to **120s**.