@robertnishihara 2019-02-15
@stephanie-wang could you confirm that these changes don't violate the spirit of the test?