@stephanie-wang 2019-02-15
Now that we have `stress_tests/test_dead_actors.py`, I don't think we even need this test in Travis anymore. What do you think?