@robertnishihara 2019-02-15
`stress_tests/test_dead_actors.py` doesn't get tested with every commit though, right? I think it'd be good to have some fault tolerance tests in Travis that get tested with every commit (though they don't need to be "stressful" so we can scale them down).